refactor: remove unwanted classes and test
