+++ failover-core/src/main/java/com/societegenerale/failover/core/observable/manifest/package-info.java
@@ -1,8 +1,8 @@
 /**
- * Manifest-based metadata extraction for the failover startup report.
+ * Manifest-based metadata extraction for the failover observable report.
  *
  * <p>Reads build-time information (version, build date) from the JAR manifest
- * or classpath resources to enrich the startup report emitted by
- * {@link com.societegenerale.failover.core.report.FailoverObserver}.
+ * or classpath resources to enrich the metrics emitted by
+ * {@link com.societegenerale.failover.core.observable.FailoverObserver}.
  */
 package com.societegenerale.failover.core.observable.manifest;
