+++ failover-core/src/main/java/com/societegenerale/failover/core/observable/package-info.java
@@ -1,9 +1,10 @@
 /**
- * Reporting and metrics publication for the failover lifecycle.
+ * Observable SPI and core implementations for the failover lifecycle.
  *
- * <p>{@link com.societegenerale.failover.core.report.FailoverObserver} emits startup, store,
- * and recover events. Built-in publishers target SLF4J and Micrometer.
- * {@link com.societegenerale.failover.core.report.CompositeObservablePublisher} fans out to
- * multiple publishers simultaneously.
+ * <p>{@link com.societegenerale.failover.core.observable.FailoverObserver} collects metrics
+ * from all registered {@code @Failover} configurations and fans them out to
+ * {@link com.societegenerale.failover.core.observable.publisher.ObservablePublisher} instances.
+ * {@link com.societegenerale.failover.core.observable.publisher.CompositeObservablePublisher}
+ * stamps a single publish timestamp and delegates to all registered publishers simultaneously.
  */
 package com.societegenerale.failover.core.observable;
