+++ failover-core/src/main/java/com/societegenerale/failover/core/observable/publisher/package-info.java
@@ -4,8 +4,8 @@
  * <p>{@link com.societegenerale.failover.core.observable.publisher.ObservablePublisher} is the
  * core SPI for publishing failover metrics to external sinks.
  * {@link com.societegenerale.failover.core.observable.publisher.AbstractObservablePublisher} provides
- * a base implementation. Built-in publishers:
- * {@link com.societegenerale.failover.core.observable.publisher.MdcLoggerObservablePublisher} (SLF4J) and
- * {@link com.societegenerale.failover.core.observable.publisher.MetricsObservablePublisher} (MDC-based).
+ * a base implementation. Default built-in publisher:
+ * {@link com.societegenerale.failover.core.observable.publisher.MdcLoggerObservablePublisher}
+ * (enriches MDC with metric attributes and logs via SLF4J).
  */
 package com.societegenerale.failover.core.observable.publisher;
