+++ failover-core/src/main/java/com/societegenerale/failover/core/observable/scanner/FailoverScannerException.java
@@ -17,7 +17,7 @@
 package com.societegenerale.failover.core.observable.scanner;
 
 /**
- * Thrown when the {@link FailoverScanner} encounters an error during classpath scanning
+ * Thrown when the {@link FailoverScanner} encounters an error during Spring context scanning
  * or detects duplicate {@code @Failover} names.
  *
  * @author Anand Manissery
@@ -31,15 +31,5 @@ public class FailoverScannerException extends RuntimeException {
      */
     public FailoverScannerException(String message) {
         super(message);
-    }
-
-    /**
-     * Creates an exception with the given detail message and cause.
-     *
-     * @param message   description of the scanner error
-     * @param throwable the underlying exception from the Reflections library
-     */
-    public FailoverScannerException(String message, Throwable throwable) {
-        super(message, throwable);
     }
 }
