--- failover-core/src/test/java/com/societegenerale/failover/core/observable/publisher/MetricsObservablePublisherTest.java
@@ -1,63 +1,0 @@
-/*
- * Copyright 2022-2023, Société Générale All rights reserved.
- *
- * Licensed under the Apache License, Version 2.0 (the "License");
- * you may not use this file except in compliance with the License.
- * You may obtain a copy of the License at
- *
- *      http://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing, software
- * distributed under the License is distributed on an "AS IS" BASIS,
- * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- * See the License for the specific language governing permissions and
- * limitations under the License.
- */
-
-package com.societegenerale.failover.core.observable.publisher;
-
-import ch.qos.logback.classic.spi.ILoggingEvent;
-import ch.qos.logback.core.Appender;
-import com.societegenerale.failover.core.observable.Metrics;
-import org.junit.jupiter.api.BeforeEach;
-import org.junit.jupiter.api.DisplayName;
-import org.junit.jupiter.api.Test;
-import org.junit.jupiter.api.extension.ExtendWith;
-import org.mockito.ArgumentCaptor;
-import org.mockito.Captor;
-import org.mockito.Mock;
-import org.mockito.junit.jupiter.MockitoExtension;
-import org.slf4j.Logger;
-import org.slf4j.LoggerFactory;
-
-import static org.assertj.core.api.Assertions.assertThat;
-import static org.mockito.Mockito.verify;
-
-@ExtendWith(MockitoExtension.class)
-class MetricsObservablePublisherTest {
-
-    private static final Metrics METRICS = Metrics.of("failover");
-
-    @Mock
-    private Appender<ILoggingEvent> appender;
-
-    @Captor
-    private ArgumentCaptor<ILoggingEvent> captor;
-
-    private MetricsObservablePublisher metricsObservablePublisher;
-
-    @BeforeEach
-    void setUp() {
-        var logger = (ch.qos.logback.classic.Logger) LoggerFactory.getLogger(Logger.ROOT_LOGGER_NAME);
-        logger.addAppender(appender);
-        metricsObservablePublisher = new MetricsObservablePublisher();
-    }
-
-    @Test
-    @DisplayName("should publish metrics")
-    void shouldPublishMetrics() {
-        metricsObservablePublisher.publish(METRICS);
-        verify(appender).doAppend(captor.capture());
-        assertThat(captor.getValue().getMDCPropertyMap()).containsAllEntriesOf(METRICS.getInfo());
-    }
-}
