--- failover-observable-micrometer/src/main/java/com/societegenerale/failover/observable/micrometer/report/package-info.java
@@ -1,5 +1,0 @@
-/**
- * Micrometer-based {@link com.societegenerale.failover.core.report.ObservablePublisher} implementations.
- * Replaces MDC-log-based fake metrics with real Micrometer counters, gauges, and timers.
- */
-package com.societegenerale.failover.observable.micrometer.report;
