+++ failover-spring-boot-autoconfigure/src/main/java/com/societegenerale/failover/configuration/FailoverMicrometerAutoConfiguration.java
@@ -19,8 +19,8 @@ package com.societegenerale.failover.configuration;
 import com.societegenerale.failover.core.expiry.FailoverExpiryExtractor;
 import com.societegenerale.failover.core.observable.publisher.ObservablePublisher;
 import com.societegenerale.failover.core.observable.scanner.FailoverScanner;
-import com.societegenerale.failover.observable.micrometer.report.FailoverMeterBinder;
-import com.societegenerale.failover.observable.micrometer.report.MicrometerObservablePublisher;
+import com.societegenerale.failover.observable.micrometer.FailoverMeterBinder;
+import com.societegenerale.failover.observable.micrometer.MicrometerObservablePublisher;
 import io.micrometer.core.instrument.MeterRegistry;
 import org.springframework.boot.autoconfigure.AutoConfiguration;
 import org.springframework.boot.autoconfigure.condition.ConditionalOnBean;
